Add user-controllable watertight scene display
The watertight scene variant is essentially a clone of all scene objects
prior to them being anchored with individual WorldAnchors. The parent of
the watertight scene objects is a separate GameObject whose transform is
kept up to date with the scene's global OriginSpatialGraphNodeId.

By default, the non-watertight scene variant (where each scene object is
anchored individually) is shown. The user can say "watertight on" to
switch the display to the watertight scene variant. (This is purely
visual and doesn't require the scene to be updated.) Saying "watertight
off" switches back to the non-watertight scene variant. This setting
sticks across updates.

The two scene variants can be visually distinguished by the text labels
shown for each scene object: In the non-watertight scene, each label has
diagnostic WorldAnchor information (like "(A1 3.2m)") attached to it; in
the watertight scene, it doesn't (because all scene objects are anchored
on the same root anchor).